rename bashi.filter in bashi.filter_base
- the module name filter could collide with the filter() function from
  the standard library
