+++ src/bashi/filter_base.py
