+++ tests/test_generate_combination_list.py
@@ -23,7 +23,7 @@ from bashi.types import (
     ParameterValueMatrix,
 )
 from bashi.globals import *  # pylint: disable=wildcard-import,unused-wildcard-import
-from bashi.filter import FilterBase
+from bashi.filter_base import FilterBase
 from bashi.row import BashiRow
